requirements.txt
apache-airflow-providers-google
pyarrow
#pyspark
pandas
openpyxl
bs4